Use provided offsets instead of heuristic to parse pilot entries
Using the presence or abscence of 40 bytes of 0xFF to determine if a
segment of memory is a valid pilot or not is unreliable - Both the
first pilot and the pilot immediately following the player are missed.

Having located seemingly canonical locations for navigating saved pilot
in the beginning of the save file, it should no longer be required.